+++ R/SEQexpand.R
@@ -9,7 +9,6 @@ SEQexpand <- function(params) {
   out <- local({
     on.exit({
       rm(list = setdiff(ls(), "out"))
-      gc()
     }, add = TRUE)
 
     # Variable pre-definition ===================================
@@ -49,7 +48,7 @@ SEQexpand <- function(params) {
     vars.sq <- unique(sub(params@indicator.squared, "", vars.sq))
     vars.kept <- c(vars, params@id, "trial", "period", "followup")
 
-    data <- DT[, list(period = Map(seq, get(params@time), table(DT[[params@id]])[.GRP] - 1)), by = eval(params@id),
+    data <- DT[, list(period = Map(seq, get(params@time), .N - 1)), by = eval(params@id),
                ][, cbind(.SD, trial = rowid(get(params@id)) - 1)
                  ][, list(period = unlist(.SD)), by = c(eval(params@id), "trial")
                    ][, followup := as.integer(seq_len(.N) - 1), by = c(eval(params@id), "trial")