+++ R/globals.R
@@ -1,0 +1,1 @@
+utils::globalVariables(c("..cols", "boot_idx", "followup", "tx_lag"))