+++ NEWS.md
@@ -1,3 +1,11 @@
+# SEQTaRget v1.3.4
+
+- Implemented some code optimizations
+  - Replace a `table()` call with data.table's `.N`
+  - Remove all `gc()` calls
+  - Use a keyed index in bootstrapping
+  - Remove some uses of `copy()`
+
 # SEQTaRget v1.3.3
 
 - Found and fixed a bug which caused excused switches to be overwritten.